requirements.txt
# 核心依赖
requests>=2.28.0
pyyaml>=6.0

# 火山引擎API SDK (使用已安装的版本，避免编译问题)
volcengine-python-sdk>=4.0.5
tos==2.8.1

# 视频处理
moviepy>=1.0.3
opencv-python>=4.6.0
pillow>=9.2.0

# 音频处理
pydub>=0.25.1

# 网络请求和重试
tenacity>=8.1.0
aiohttp>=3.8.0

# 日志和工具
loguru>=0.6.0
tqdm>=4.64.0
python-dotenv>=1.0.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 开发工具
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0